[H2BLB][FastISel] Support input arguments in registers
This patch adds the support of input arguments when they are passed
through registers.
This is an example of how FastISel can deal with input arguments and
the take away is that you are on your own (i.e., this is pure C++).

Check the isel-only-abi.ll file to see the intermediate lowering step.