[H2BLB] This hook up GISel for our target
This commit implements a skeleton of the GlobalISel pipeline for our target.
The GISel pipeline consists in four mandatory phases:
1. IRTranslator, which requires a XXXCallLowering instance
2. Legalizer, which requires a XXXLegalizerInfo instance
3. RegisterBankSelect, which requires a XXXRegisterBankInfo
4. InstructionSelect, which requires a XXXInstructionSelector

In this commit we only provide the H2BLBCallLowering and test the pipeline
until after IRTranslator.

The implementation of H2BLBCallLowering may look complex, but this is
just a reimplementation of everything that has been done in SDISel so far.
Therefore, you should be able to follow the general structure.