[H2BLB][ISel] Select widen mul from IR constructs
This commit teaches ISel how to selection widen mul instructions even
when we don't use the intrinsic.
One can see here how a pattern with sub-dag can be written.

This patch is not enough for GISel because we didn't teach it how to
let G_MUL pass through legalization.

It is enough for SDISel though, because by default every operation is
legal as soon as the type is legal.