[H2BLB][SDISel] Add a custom dag combine
This commit demonstrates how to use DAG combines.

The combine takes mul(sext,sext) and produce widen mul.
This is yet another way patterns can be matched throughout the selector.

Realistically, we don't need 3 different implementations (custom combine,
custom legalization, selection dag pattern), but this illustrates the
different options we have.

Looking at select-widen-mul.ll, one can see the different options in
action by running the test with -debug-only=isel.

For mode 0, the instruction is selected during instruction selection.
For mode 1, the instruction is lowered to a target specific SDNode
during legalization and that node is then selected to the appropriate
instruction.
For mode 2, the instruction is lowered to a target specific SDNode during
the first DAG combine run (before legalization), then that node is selected
appropriately during instruction selection.

This commit builds on top of the previous commits to offer the different
paths of lowering (i.e., this commit only adds mode 2).