[H2BLB] Enable the Machine scheduler
This commit overrides the target hook used to enable the machine
scheduler.

The default vs. generic checks are identifical for now because we
didn't hook up any scheduling model for our target yet.

Note that the second override (`enableMachineSchedDefaultSched`) is
here to keep the default SDISel scheduler in place, just because we
didn't want to create noise in the tests.