[H2BLB][FastISel] Add support for return values
This patch teaches how to select return instruction with FastISel
when the return has a value to return. In other words, before we
supported `ret void` and now we support `ret ty value` where `ty`
is a simple type.

This code leverages the fastEmit_i function that is automatically
generated by TableGen. The call to this function is hidden behind
the `getRegForValue`. This method looks at a map of which values
are available in which registers and when the value is not available
and not produced by an instruction (for instance, when this is a
constant), it produces it on the spot.