[H2BLB][Legalizer] Add the rules for build_vector, merge, and unmerge
This commit tells the legalizer that the build_vector, merge, and unmerge
instructions are legal when the input (resp. output for unmerge) operands
are 16-bit and the output operand (resp. input) is 32-bit.

The idea is this represents what we can do we our subregister indices:
low(d0) == r0 and high(d0) == r1.