[H2BLB] Buff up the calling convention support
Move the calling convention information in its own file to be
able to share the logic between SDISel and FastISel.

Add a calling convention (called common) for lowering arguments
and returns. The argument part is empty for now.

The use of the `Entry` property on the calling conventions (in the
.td files) is what make it possible to reuse them between FastISel
and SDISel.