[H2BLB][SDISel] Add truncstore legalization rule
This commits teaches the legalization phase in SDISel how to deal with
truncstores that take a 32-bit value in and truncate it to a 16-bit on
store.

The `Expand` action means that the legalizer will take the input SDNode
and replace it with a sequence of semantically equivalent SDNodes.

For instance, the expanded sequence of a truncstore is truncate followed
by store.