[H2BLB][GISel] Hook up the custom inserter for RET_PSEUDO
This commit replaces how RETURN instructions are lowered in GISel by
introducing a pseudo instruction that is used until finalize-isel.

The reason we introduce this pseudo is to avoid to expose the implicit use of
R0 too early. Specifically, R0 (our link register) is indeed used by our RETURN
instruction but its live-range will only be materialized during the
finalize-isel pass. Therefore, until we reach this pass, the live-range of this
register will be incorrect (a use without a def) and will be flagged as such
by the machine verifier.

This problem prevented enabling the machine verifier, hence this fix.