[H2BLB][GISel] Select add ptr
In GlobalISel adds on pointer types use a different opcode.
Therefore the regular add pattern doesn't work.

In theory it is possible to simply add a pattern that match ptr add
to our regular add.
In practice, this is not possible in our case because this creates a
conflict in the FastISel TableGen backend. Indeed, this TableGen
backend sees two different patterns for an add instructions (add and
ptradd are the same thing outside of GISel world) and abort because
it cannot handle duplicated patterns.

Therefore, we have to instroduce a new pseudo instruction to anchor
our selection pattern on something else and tell the FastISel backend
to ignore this something else.

The custom inserter is here to get rid of this pseudo instruction.