[H2BLB] Don't run finalizeLowering twice
The instruction select pass in GlobalISel already calls the
TargetLowering::finalizeLowering method.

This method is also called by the finalize-isel pass.
In this method we fix the liveness of our link register and we don't want
to do it twice.
Therefore check if the code is already selected (property set by the GISel
selector) before fixing the liveness.