[H2BLB][GISel] Match non-SP addressing mode
This commit teaches GISel how to match non-SP addressing mode.
More specifically, this patch implements the matching of the equivalent
SDISel complex pattern and tells TableGen that this implementation is
indeed equivalent to the SDISel implementation (see GIComplexPatternEquiv).

This allows the TableGen pattern importer to grab the SDISel patterns that
use that complex pattern.