[H2BLB][ISel] Add F32 support via libcalls
This commit adds the support for fadd f32 operations.
It includes the changes in legalization info and the patterns for ISel.

The legalization changes are the most interesting as they demonstrate how
to use software emulation for floating point types using the libcall lowering.

The GlobalISel part of the legalizer is also modified for loads, so that
we scalarize them.