[H2BLB] Placeholder for the upcoming backend
Add the bare minimum to be able to build the H2BLB.

This means featuring:
- a LLVMInitializeH2BLBTargetMC that is being called from
  InitializeAllTargetMCs
- a LLVMInitializeH2BLBTargetInfo that is being called from
  InitializeAllTargetInfos
- a LLVMInitializeH2BLBTarget that is being called from
  InitializeAllTargets

The actual structure of the directories set us on the right foot
for what is coming.

Thanks to this commit, we can use H2BLB in -DLLVM_TARGETS_TO_BUILD.
That being said, it doesn't do anything useful.