[H2BLB] Set the default scheduling direction as top-down
This commit demonstrates how to change the default scheduling direction
on a per sub-target basis.

In the sched.mir test, we can see that this the top-down approach the
latency of the load is filed with the ADD instructions instead of the
COPY instruction.

With this direction, we can now see the difference in the default
scheduling model and the scheduling model of our `generic` processor.
The default model has a load latency of 4 and our `generic` processor
a latency of 3, which is reflected in the number of instructions
that are inserted between the load instruction (`LDR16`) and its
consumer, the widening mul instruction (`WSMUL`).