[H2BLB] Implement the CALL instruction
This patch continues with augmenting the ISA of the H2BLB target as
we need it.
The CALL instruction consists in this case of just a jump to an address
represented by an 11-bit immediate value.

To be able to print this value with the MC layer, we have to update:
- Our AsmPrinter to teach it how to handle symbols
- Our MCCodeEmitter to teach how to add a fix-up for the call target

A fix-up is an indication left by the assembler to the linker that tells
the linker that it needs to patch the final binary with something.
In this specific case, the fix-up would tell the liker to replace the 11-bit
of the immediate value with the actual address (PC-relative) of the related
symbol.
The fix-up is currently not emitted because we didn't connect a target
specific MCAsmBackend yet.