[H2BLB] Implement the lowering of the stack
This is a primitive lowering of the stack as in it doesn't support
saving and restoring callee saved registers.

The prologue consists in decreasing the value of SP and the epilogue
increasing it back.
Therefore our stack grows downward.

We do not support any variable size objects, like dynamic allocas.
For this we would need a frame pointer and given how few registers
our target has we do not want to freeze one more register for this.