[H2BLB][GISel] Select COPY-like instructions
This commit teaches the selector how to produce the non-generic form
of the COPY and PHI instructions.

These instructions are special because they are not pre-selection opcodes,
like all the G_xxx opcodes, and yet, the selector needs to change them.
Indeed, non-pre-selection opcodes are expected to be already selected.
However PHI and COPY instructions are pseudo instructions that can be used
both before and after instruction selection and must go through selection
to make sure they comply with the machine IR requirements.

Indeed these instructions may still hold generic virtual
register and the actually register class needs to be set.