[H2BLB][SDISel] Show how to do custom legalization
This commit lowers MUL operation via a custom legalizer hook that
produces the WINDEN_MUL operations directly.
To do that we must introduce new SDNodes and select them.

When doing custom legalization, if the custom legalization fails
SDISel tries to use the Expand action next and if this still doesn't
work it uses the libcall action.

Note that there is no new test, because it all happens within SDISel
and since we already knew how to general widen mul, the difference
is not observable in the output (or input).
The way this can be observed is by looking at the debug output of
SDISel and see that the widen mul nodes appears during legalization
(e.g., on select-widen-mul.ll)