[H2BLB][ISel] Implement finalizeLowering
This commit adds the proper code to save and restore R0, our link register,
at the beginning and the end of each MachineFunction.

This allows us to enable the machine verifier now that R0 has the proper
live-range.

Note that this is a way to handle the link register. A popular other way
is to mark it reserved and handle the save/restore during the frame lowering.

The version in this commit takes a different approach and handles the save
and restore of R0 through explicit copies because these copies can be coalesced
for free when the compiled function does not have any calls in it.