[H2BLB][RegBankInfo] Add register mapping
This commit teaches our RegisterBankInfo about the value mapping for our
instructions.

Since we only have one register bank, the gist of this change is about
providing static tables for the most common cases and resort on the
dynamic table for the others.

The static table can represent operations with up to 3 operands (hence
including the definition) for both 16 and 32-bit instruction.