[H2BLB] Add some scheduling event
This commit demonstrates how to use two of the constructs to add scheduling
events to an instruction description.

The first one is the Sched class which can be appended to an instruction
to specify its scheduling event.

The second one consists in setting directly the SchedRW field in the instruction
description.

The semantic of the list of events are interpreted the same (see the comments
on the Sched class to understand them), the only thing that changes is how the
SchedRW field is ultimately set.
For the first method, the SchedRW is implicitly overrided because the Sched
class also defines a SchedRW field and this version just stomp on the previous
one.
For the second method, we set the field directly.

These events don't carry any meaning beyond read or write. Their effects
on the scheduling is micro-architecture specific and as such depends on the
scheduling model.